Fixing user flow issue (TODO 1/2/3/4)
