+++ src/components/MenuItemPreviewCard.jsx
@@ -30,7 +30,7 @@ const MenuItemPreviewCard = ({ className, onClickMenuItem, ...props }) => (
         <span className="text-primary font-heading"> {props.name} </span>
       </div>
       <div>
-        <span className="text-primary font-paragraph"> {props.price} </span>
+        <span className="text-primary font-paragraph"> ${props.price} </span>
       </div>
     </div>
   </Card>
