+++ src/routes/UserHome.jsx
@@ -9,29 +9,10 @@ const UserHome = () => {
     const location = useLocation();
     const fromJoinSuccess = location.state?.fromJoinSuccess;
 
-    const handleJoin = () => {
-        navigate('/user-home/join-challenge');
-    }
 
-    const handleJoinSuccess = () => {
-        navigate('/user-home/join-challange-success');
-    }
 
     return (
         <>
-            <div className="py-2">
-                <ChallengeModal/>
-                {!fromJoinSuccess && (
-                    <Link onClick={handleJoin}>
-                        JOIN ZERO-WASTE CHALLENGE!
-                    </Link>
-                )}
-                {fromJoinSuccess && (
-                    <Link onClick={handleJoinSuccess}>
-                        YOU HAVE JOINED THE CHALLENGE!
-                    </Link>
-                )}
-            </div>
             <MenuItemsPage />
         </>
     );
